Configure-able player name drawing
Currently, hottracked player names are drawn on top of crosshair
together with distance to them. Center of screen is important region and
thus it should be configure-able.  This allows you to disable drawing
name on the whole or just disable drawing the distance. Also possible to
position the name wherever you want.