Updated spectator behavior.
When spectating, primary and secondary fire (default left and right click)
now go to the next and previous player, respectively. I did this so that the
spectator doesn't have to cycle through all the players again when he/she
"overshoots" by spamming click too much.

The button to cycle through modes is now jump (default spacebar). it goes
from "free", to "follow", to "first person". If the player is not on the spectator
team, then skips through free automatically.

Lastly, I moved the "Following {playername}" text to the top of the screen.
I also added the player's ID#. This is useful if you're spectating a possible
cheater and you need to report its ID#.