Revert "Run update-pot.sh (Main.as not found?)"
This reverts commit 529d652ff6f7b876545f24170a6d13f30a7bf487.

I am not sure if I should include that commit, so better not to.
If executing update-pot.sh is a nescessary step than beter leave
it to someone else, as when I executed the script it said that
Main.as was not found, so maybe something was broken in that commit.