Adjust the gain of hit feedback sound
Fixes #652.

The sound gain is now constant and does not change regardless of the
distance. This imitates the behavior often seen in other first-person
shooters.

Switching from "Play" to "PlayLocal" only affects the way the gain is
computed since "HeadshotFeedback.opus" is a stereo sound file and there-
fore not spatialized by any known audio backends.

This also introduces a config variable named "cg_hitFeedbackSoundGain"
that allows users to adjust the volume of the hit feedback sound.