Reset intels and team scores after receiving winning IntelCapture
This problem was originally reported in:
<https://github.com/piqueserver/piqueserver/issues/586>

Upon receiving `IntelCapture` with `winning = 1`, the voxlap client
resets both team scores and the holding status of both teams' intels,
but OpenSpades didn't do that, resulting in a weird behavior described
in piqueserver/piqueserver#586. This commit changes OpenSpades' behavior
to be congruent with the voxlap client's behavior.

Fixes #890.