feat(draw): gate `GLFogFilter2` by `r_radiosity` and `r_temporalAA`
I intend the volumetric fog to receive GI. To simplify the
implementation, I decided not to support other cases.

I'd like to exploit temporal AA to reduce the sample count per pixel
by dithered sampling. It would perplex the users if the unfiltered
dither pattern was visible because temporal AA was turned off.