Player.cpp: Disable recoil if not local player
This commit completely disables any sign of recoil if
the packet is coming from a player that is not local
player.

We want to apply recoil to local player weapon
but we do not want to apply it to others as
we just want to display what was given to us
(Their client already adds recoil. If not they
are cheating)